+++ pySDC/projects/Second_orderSDC/penningtrap_run_Hamiltonian_error.py
@@ -11,7 +11,7 @@ from pySDC.implementations.problem_classes.PenningTrap_3D import penningtrap
 from pySDC.implementations.sweeper_classes.boris_2nd_order import boris_2nd_order
 from pySDC.projects.Second_orderSDC.penningtrap_HookClass import particles_output
 from pySDC.implementations.sweeper_classes.Runge_Kutta_Nystrom import RKN
-from pySDC.projects.Second_orderSDC.penningtrap_Simulation import fixed_plot_params
+from pySDC.projects.Second_orderSDC.plot_helper import set_fixed_plot_params
 
 
 def main(dt, tend, maxiter, M, sweeper):  # pragma: no cover
@@ -97,9 +97,9 @@ def plot_Hamiltonian_error(K, M, dt):  # pragma: no cover
         M: number of quadrature nodes
         dt: time step
     """
-    fixed_plot_params()
+    set_fixed_plot_params()
     # Define final time
-    time = 1e6
+    time = 1e+6
     tn = dt
     # Find time nodes
     t = np.arange(0, time + tn, tn)
